Fix remaining tests
